Changed App.js--broke it to components
