+++ frontend/frontend/src/index.css
@@ -32,7 +32,6 @@ h1 {
 
 .app:before {
   content: '';
-  background: url('./assets/sunset.jpg') no-repeat center center/cover;
   position: absolute;
   width: 100%;
   height: 100%;
